+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next
 
+* [PR 532] Add app tracking transparency
+
 ## [2.8.0] 02/09/2021
 
 * [PR 530] Add custom sign message via integrator